Prevent operation specs unintentionally passing
The final set of specs for 'GET /:operation/:number1/:number2' do not
include a test for a correct response status.  Without this test if the
error page has any of the numbers that are the expected result of the
operations the tests will pass.

Adding 'expect(last_response.status).to eq(200)' to each of the specs
prevents this error.